KNN for Regression
Task 01:
Your objective is to implement the K-Nearest Neighbors (KNN) algorithm
with k=3. Use this algorithm to predict the income value, with the given
input value X=22.
Final Task: Generate an Excel spreadsheet for a provided dataset.
•
Use Python programming to validate the KNN calculations.
•
Additionally, compare the predicted outcomes of the KNN algorithm
with Linear Regression results.